python-azure-0auth-sample/requirements.txt
cryptography==3.1
Flask==1.1.2
PyJWT==1.7.1
python-dotenv==0.14.0
requests==2.24.0
gunicorn